refactor(rust): move refactored crates into main crate
