+++ implementations/rust/Cargo.toml
@@ -13,10 +13,12 @@ members = [
     "queue_topic",
     "transport",
     "node",
-#    "worker",
     "c/generate_bindings",
     "c/bindings",
     "c/rust_memory",
+    "message_router",
+    "tcp_manager",
+    "worker_manager",
 ]
 
 default-members = [
@@ -32,7 +34,9 @@ default-members = [
     "queue_topic",
     "transport",
     "node",
-#    "worker",
     "c/bindings",
     "c/rust_memory",
+    "message_router",
+    "tcp_manager",
+    "worker_manager",
 ]
