+++ implementations/rust/_node/Cargo.toml
@@ -1,24 +1,32 @@
 [package]
 name = "ockam-node"
 version = "0.1.0"
-authors = ["Robin Budd <robin@ockam.io>"]
+authors = ["Ockam Developers"]
 edition = "2018"
+
+[[bin]]
+name = "node"
+path = "src/bin/main.rs"
+test = false
+bench = false
 
 [lib]
 crate-type = ["rlib"]
 
-[profile.release]
-lto = true
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
 ockam = { "version" = "0.1", path = "../ockam" }
 ockam-router = { version = "0.1", path = "../router" }
 ockam-transport = { version = "0.1", path = "../transport"}
+ockam-message = { version = "0.1", path = "../message"}
 ockam-common = { version = "0.1", path = "../common"}
 ockam-kex = { version = "0.1", path = "../kex" }
-ockam-vault-software = { version = "0.1", path = "../vault/software" }
+ockam-vault = { version = "0.1", path = "../vault" }
 ockam-worker = { version = "0.1", path = "../worker"}
+ockam-system = { version = "0.1", path = "../system" }
 
 hex = "0.4.2"
 structopt = { version = "0.3.20", default-features = false }
 rand = "0.7"
+
