+++ implementations/rust/_node/src/blast_worker.rs
