+++ implementations/rust/no_std_traits/src/lib.rs
