--- implementations/rust/refactor_prototype/Cargo.toml
@@ -1,19 +1,0 @@
-[workspace]
-
-members = [
-    "message_router",
-    "no_std_traits",
-    "node",
-    "ockam",
-    "tcp_manager",
-    "worker_manager",
-]
-
-default-members = [
-    "message_router",
-    "no_std_traits",
-    "node",
-    "ockam",
-    "tcp_manager",
-    "worker_manager",
-]
