--- implementations/rust/refactor_prototype/node/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "ockam-node"
-version = "0.1.0"
-authors = ["Robin Budd <robin@ockam.io>"]
-edition = "2018"
-
-[lib]
-crate-type = ["rlib"]
-
-[profile.release]
-lto = true
-
-[dependencies]
-ockam-message = { version = "0.1", path = "../../message" }
-ockam-message-router = { version = "0.1", path = "../message_router" }
-ockam-no-std-traits = { version = "0.1", path = "../no_std_traits" }
-ockam-tcp-manager = {version = "0.1", path = "../tcp_manager"}
-ockam-worker-manager = {version = "0.1", path = "../worker_manager"}
-libc = "0.2.80"
-libc-print = "0.1.14"
