--- implementations/rust/refactor_prototype/ockam/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "ockam"
-version = "0.1.0"
-authors = ["Ockam Developers"]
-edition = "2018"
-
-[dependencies]
-ockam-message = { version = "0.1", path = "../../message", default-features = false }
-ockam-message-router = { version = "0.1", path = "../message_router", default-features = false }
-ockam-no-std-traits = { version = "0.1", path = "../no_std_traits", default-features = false }
-ockam-node = { version = "0.1", path = "../node", default-features = false }
-ockam-worker-manager = { version = "0.1", path = "../worker_manager", default-features = false }
-ockam-tcp-manager = { version = "0.1", path = "../tcp_manager", default-features = false }
