--- implementations/rust/refactor_prototype/ockam/src/lib.rs
@@ -1,6 +1,0 @@
-pub use ockam_message as message;
-pub use ockam_message_router as message_router;
-pub use ockam_no_std_traits as no_std_traits;
-pub use ockam_node as node;
-pub use ockam_tcp_manager as tcp_manager;
-pub use ockam_worker_manager as worker_manager;
