+++ implementations/rust/tcp_manager/Cargo.toml
@@ -14,6 +14,6 @@ lto = true
 
 [dependencies]
 ockam-no-std-traits = { version = "0.1", path = "../no_std_traits" }
-ockam-message = { version = "0.1", path = "../../message" }
+ockam-message = { version = "0.1", path = "../message" }
 libc = "0.2.80"
 libc-print = "0.1.14"
