+++ implementations/rust/tcp_manager/src/lib.rs
