--- implementations/rust/worker/src/lib.rs
@@ -1,3 +1,0 @@
-#![allow(dead_code)]
-pub mod worker;
-pub mod worker_manager;
