+++ implementations/rust/worker_manager/src/lib.rs
